fix(sim): PR #580 deferred items — rope_scaling extraction, MaxModelLen int64, tests, docs (#606)
Complete 7 deferred hardening items from issue #580 (PR #587 handoff):

1. Extract applyRopeScaling as a pure function with 26 table-driven test
   cases covering blacklist (su/longrope/llama3), mrope fall-through,
   gemma3 substring match (handles text_config pivot), yarn original base,
   overflow guards, NaN/Inf defense, degenerate inputs.

2. Change MaxModelLen from int to int64 for consistency with ProgressIndex,
   TotalKVBlocks, BlockSizeTokens. Updates 6 type sites, removes redundant
   int64() casts, adds int64() widening at EnqueueRequest comparison sites.

3. Add cluster-mode MaxModelLen drop test (BC-6): Guard 1a (input >= limit)
   and Guard 1b (input + budget > limit), INV-1 conservation, inFlightRequests
   drain, Metrics.Requests map cleanup.

4. Add chunked prefill + MaxModelLen interaction test (BC-7): verifies no
   spurious force-completion during multi-chunk prefill (TotalOutputTokens=49,
   LengthCappedRequests=0, TTFT recorded).

5. Add glossary entries for MaxModelLen and Oracle Knowledge Boundary (INV-9).

6. Refine rope_scaling documentation with explicit blacklist details.

7. Fix pre-existing gemma3 bug: ParseHFConfig's text_config pivot overwrites
   model_type from "gemma3" to "gemma3_text", making the exact-match check
   dead code. Changed to strings.Contains to match vLLM's substring semantics.

Related to #580. Discovered issues: #602, #603, #604, #605.

+++ sim/config.go
@@ -79,14 +79,14 @@ type ModelHardwareConfig struct {
 	GPU         string        // GPU type (e.g., "H100")
 	TP          int           // tensor parallelism degree
 	Backend     string        // latency model backend: "" or "blackbox" (default), "roofline", "crossmodel"
-	MaxModelLen int           // max total sequence length (input + output); 0 = unlimited (mirrors vLLM --max-model-len)
+	MaxModelLen int64         // max total sequence length (input + output); 0 = unlimited (mirrors vLLM --max-model-len)
 }
 
 // NewModelHardwareConfig creates a ModelHardwareConfig with all fields explicitly set.
 // This is the canonical constructor — all construction sites must use it (R4).
 // Parameter order matches struct field order.
 func NewModelHardwareConfig(modelConfig ModelConfig, hwConfig HardwareCalib,
-	model, gpu string, tp int, backend string, maxModelLen int) ModelHardwareConfig {
+	model, gpu string, tp int, backend string, maxModelLen int64) ModelHardwareConfig {
 	if maxModelLen < 0 {
 		panic(fmt.Sprintf("NewModelHardwareConfig: MaxModelLen must be >= 0, got %d", maxModelLen))
 	}
